Move shard checklist config to a folder
